# Conflicts:
#   app/src/main/java/com/example/myapplication/MainActivity.kt

+++ app/src/main/java/com/example/myapplication/room/TransactionEntity.kt
@@ -27,7 +27,7 @@ data class TransactionEntity(
     val longitude: Double,
 
     @ColumnInfo(name = "nominal")
-    val nominal: Int,
+    val nominal: Double,
 
     @ColumnInfo(name = "category")
     val category: CategoryEnum
